Moved module to a different file
Moved module to a different file and created a main script.
